Remove code and tests for pydantic v1 (probably not all).
