--- src/pydantic_yaml/examples/v1/__init__.py
@@ -1,1 +1,0 @@
-"""Examples for Pydantic v1."""
