--- src/pydantic_yaml/examples/v2/__init__.py
@@ -1,1 +1,0 @@
-"""Examples for Pydantic v2."""
