+++ src/test/test_non_class_types.py
@@ -1,10 +1,10 @@
 """Regression tests for when 'type' to parse as is not a class."""
 
-import sys
 from typing import Union
 
-from pydantic_yaml import parse_yaml_raw_as
 from pydantic import BaseModel
+
+from pydantic_yaml import parse_yaml_raw_as
 
 
 class MyStrModel(BaseModel):
@@ -33,17 +33,15 @@ def test_non_class_types_union() -> None:
     assert isinstance(inst2, MyListModel)
 
 
-if sys.version_info > (3, 10, 0):
-
-    def test_non_class_types_uniontype() -> None:
-        """Test non-class types."""
-        inst1: MyStrModel | MyListModel = parse_yaml_raw_as(  # type: ignore
-            MyStrModel | MyListModel,  # type: ignore
-            "my_prop: my_yaml_string",
-        )
-        assert isinstance(inst1, MyStrModel)
-        inst2: MyStrModel | MyListModel = parse_yaml_raw_as(  # type: ignore
-            MyStrModel | MyListModel,  # type: ignore
-            "my_prop: [my_yaml_string]",
-        )
-        assert isinstance(inst2, MyListModel)
+def test_non_class_types_uniontype() -> None:
+    """Test non-class types."""
+    inst1: MyStrModel | MyListModel = parse_yaml_raw_as(  # type: ignore
+        MyStrModel | MyListModel,  # type: ignore
+        "my_prop: my_yaml_string",
+    )
+    assert isinstance(inst1, MyStrModel)
+    inst2: MyStrModel | MyListModel = parse_yaml_raw_as(  # type: ignore
+        MyStrModel | MyListModel,  # type: ignore
+        "my_prop: [my_yaml_string]",
+    )
+    assert isinstance(inst2, MyListModel)
